shorter trainer test

+++ src/pytorch_metric_learning/utils/common_functions.py
@@ -223,7 +223,7 @@ def get_labels_to_indices(labels):
     for i, label in enumerate(labels):
         labels_to_indices[label].append(i)
     for k, v in labels_to_indices.items():
-        labels_to_indices[k] = np.array(v, dtype=np.int)
+        labels_to_indices[k] = np.array(v, dtype=int)
     return labels_to_indices
 
 
@@ -265,7 +265,7 @@ class LabelMapper:
             return labels
         else:
             return np.array(
-                [self.label_map[hierarchy_level][x] for x in labels], dtype=np.int
+                [self.label_map[hierarchy_level][x] for x in labels], dtype=int
             )
